pyproject.toml
[project]
name = "hydrus"
version = "v645"
description = "A personal booru-style media tagger"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # requirements_core.txt
    "beautifulsoup4>=4.0.0",
    "cbor2>=5.6.5",
    "chardet>=3.0.4",
    "cryptography>=44.0.0",
    "dateparser==1.2.1",
    "html5lib>=1.0.1",
    "lxml>=4.5.0",
    "lz4>=3.0.0",
    "olefile>=0.47",
    "pillow>=10.0.1",
    "pillow-heif>=0.12.0",
    "pillow-jxl-plugin>=1.3.0",
    "psutil>=5.0.0",
    "pympler>=1.1",
    "pyopenssl>=19.1.0",
    "pysocks>=1.7.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=5.0.0",
    "send2trash>=1.5.0",
    "service-identity>=18.1.0",
    "show-in-file-manager>=1.1.5",
    "twisted[http2,tls]>=20.3.0",
    "numpy<=2.3.1",
    "requests==2.32.4",
    # requirements_macos.txt
    "pyobjc-core>=10.1 ; sys_platform == 'darwin'",
    "pyobjc-framework-cocoa>=10.1 ; sys_platform == 'darwin'",
    "pyobjc-framework-quartz>=10.1 ; sys_platform == 'darwin'",
    # requirements_windows.txt
    "pywin32 ; sys_platform == 'win32'",
]

[tool.uv]
default-groups = [ "mpv-new", "opencv-new", "other-normal", "qt6-new" ]
conflicts = [
    [
        { group = "mpv-new" },
        { group = "mpv-test" },
        { group = "mpv-old" },
    ],
    [
        { group = "opencv-new" },
        { group = "opencv-test" },
        { group = "opencv-old" },
    ],
    [
        { group = "other-future" },
        { group = "other-normal" },
    ],
    [
        { group = "qt6-new" },
        { group = "qt6-new-pyqt6" },
        { group = "qt6-older" },
        { group = "qt6-test" },
    ]
]

[dependency-groups]
mpv-new = [
    "mpv==1.0.7",
]
mpv-test = [
    "mpv==1.0.7",
]
opencv-new = [
    "opencv-python-headless==4.11.0.86",
]
opencv-old = [
    "opencv-python-headless==4.8.1.78",
]
opencv-test = [
    "opencv-python-headless==4.12.0.88",
]
other-future = [
]
other-normal = [
]
qt6-new = [
    "pyside6==6.8.3",
    "qtpy==2.4.3",
]
qt6-new-pyqt6 = [
    "pyqt6",
    "pyqt6-charts",
    "qtpy",
]
qt6-older = [
    "pyside6==6.3.1",
    "qtpy==2.3.1",
]
qt6-test = [
    "pyside6==6.9.3",
    "qtpy==2.4.3",
]
